// frontend/src/api/model/rafflePrizeRead.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Map the Mess API
 * Backend for the community litter reporting platform
 * OpenAPI spec version: dev
 */
import type { RafflePrizeImageRead } from './rafflePrizeImageRead';
import type { RafflePrizeWinner } from './rafflePrizeWinner';

export interface RafflePrizeRead {
  id: string;
  raffle_id: string;
  title: string;
  description: string;
  position: number;
  winner_user_id?: string | null;
  winner?: RafflePrizeWinner | null;
  images?: RafflePrizeImageRead[];
  created_at: string;
}